ai-api/requirement.txt
pyarrow
pandas
numpy
sentence-transformers
chromadb
sentencepiece 
tiktoken
mysql-connector-python
python-dotenv
playwright
feedparser 
beautifulsoup4
google-genai
serpapi
pillow
tensorflow